Initial WASM compilation support (not execution support) Added rank search and location, but keeping track of the ranks is still WIP

+++ Platform/CppDummy.hpp
@@ -11,7 +11,7 @@
 }
 
 void operator delete(void* ptr) noexcept {}
-inline void* operator new(size_t, void* ptr) noexcept {
+inline void* operator new(unsigned long, void* ptr) noexcept {
     return ptr;
 }
